Add initial setup for chat application
This commit adds all the necessary setup files for a chat application. It includes Spring Boot application and test classes, application properties, Bootstrap HMTL, CSS and Javascript files, as well as configurations for multiple IDEs and build tools in .gitignore. The setup uses SockJS and STOMP to facilitate communication between the server and its clients represented as a WebSocket fallback.